requirements.txt
Flask
python-dotenv
PyPDF2
google-genai